requirements.txt
Django==3.2.4
djangorestframework==3.12.4
python-dateutil==2.8.1
psycopg2-binary==2.8.5

